zip_list {zip}	R Documentation

List Files in a 'zip' Archive

Description

List Files in a 'zip' Archive

Usage

zip_list(zipfile)

Arguments

zipfile
Path to an existing ZIP file.

Details

Note that crc32 is formatted using as.hexmode(). offset refers to the start of the local zip header for each entry. Following the approach of seek() it is stored as a numeric rather than an integer vector and can therefore represent values up to 2^53-1 (9 PB).

Value

A data frame with columns: filename, compressed_size, uncompressed_size, timestamp, permissions, crc32 and offset.

[Package zip version 2.2.0 Index]